migrate page attachment code to page attachment file
